HELP ME
I thought I had this, but I was wrong. I got the links to show in the
XHTML output, but I cant get the links to point to each individual poem.
I also tried to get the table of contents to rearrange alphabetically,
but cant remember how to check the lines after the numbers. Are Andrew's
notes not on GitHub? Heelllpppppp MMEeeeeeeee!!!!!!!!!!!!!!!